Create Jar
This is the jar you use to launch the game. Remember that the "MAP" folder must be on your machine in order to launch the program.

Example: Place both the Map folder in its entirety onto your desktop and place the jar folder onto your desktop then double click the jar file to launch.