Create Executable_file
This is the executable file for the game. It is recommended that you use this version since it runs much faster.

Place both  the Map folder and the executable folder in itheir entirety onto your desktop. Then you can double click the "ConnectUs.exe" that is located in the executable folder and it should run.

Enjoy.